+++ app/src/main/java/com/battlelancer/seriesguide/lists/SgListFragment.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2012-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.lists
 
@@ -109,8 +109,8 @@ class SgListFragment : Fragment() {
         model.updateQuery()
     }
 
-    private val onItemClickListener: SgListItemAdapter.OnItemClickListener =
-        object : SgListItemAdapter.OnItemClickListener {
+    private val onItemClickListener: SgListItemViewHolder.OnItemClickListener =
+        object : SgListItemViewHolder.OnItemClickListener {
             override fun onItemClick(anchor: View, item: SgListItemWithDetails) {
                 Utils.startActivityWithAnimation(
                     requireActivity(),
@@ -123,13 +123,18 @@ class SgListFragment : Fragment() {
                 val popupMenu = PopupMenu(anchor.context, anchor)
                 popupMenu.inflate(R.menu.lists_popup_menu)
                 val menu = popupMenu.menu
-                menu.findItem(R.id.menu_action_lists_favorites_add).isVisible = !item.favorite
-                menu.findItem(R.id.menu_action_lists_favorites_remove).isVisible = item.favorite
+                // Hide some options that only make sense for shows (for legacy non-show items)
+                val isShow =
+                    item.type == ListItemTypes.TMDB_SHOW || item.type == ListItemTypes.TVDB_SHOW
+                menu.findItem(R.id.menu_action_lists_watched_next).isVisible = isShow
+                menu.findItem(R.id.menu_action_lists_favorites_add).isVisible =
+                    isShow && !item.favorite
+                menu.findItem(R.id.menu_action_lists_favorites_remove).isVisible =
+                    isShow && item.favorite
                 popupMenu.setOnMenuItemClickListener(
                     PopupMenuItemClickListener(
                         requireContext(), parentFragmentManager,
-                        item.listItemId, item.showId,
-                        item.nextEpisode?.toLongOrNull() ?: 0
+                        item.listItemId, item.showId, item.nextEpisodeId
                     )
                 )
                 // Hide manage lists option for legacy show items, only allow removal.
@@ -139,10 +144,8 @@ class SgListFragment : Fragment() {
                 popupMenu.show()
             }
 
-
-            override fun onFavoriteClick(showId: Long, isFavorite: Boolean) {
-                SgApp.getServicesComponent(requireContext()).showTools()
-                    .storeIsFavorite(showId, isFavorite)
+            override fun onItemSetWatchedClick(item: SgListItemWithDetails) {
+                EpisodeTools.episodeWatchedIfNotZero(context, item.nextEpisodeId)
             }
         }
 
@@ -162,26 +165,32 @@ class SgListFragment : Fragment() {
                     EpisodeTools.episodeWatchedIfNotZero(context, nextEpisodeId)
                     return true
                 }
+
                 R.id.menu_action_lists_favorites_add -> {
                     showTools.storeIsFavorite(showId, true)
                     return true
                 }
+
                 R.id.menu_action_lists_favorites_remove -> {
                     showTools.storeIsFavorite(showId, false)
                     return true
                 }
+
                 R.id.menu_action_lists_manage -> {
                     ManageListsDialogFragment.show(fragmentManager, showId)
                     return true
                 }
+
                 R.id.menu_action_lists_update -> {
                     ShowSync.triggerDeltaSync(context, showId)
                     return true
                 }
+
                 R.id.menu_action_lists_remove -> {
                     ListsTools.removeListItem(context, itemId)
                     return true
                 }
+
                 else -> return false
             }
         }